breakdown menu functions
