+++ include/menu.h
@@ -1,0 +1,10 @@
+#ifndef MENU_H
+#define MENU_H
+
+#include <vector>
+#include <string>
+
+int menu(const char* msg, const std::vector<std::string>& options);
+bool confirm(const char* question, const char* message = nullptr);
+
+#endif // MENU_H
